Add perfetto library fork
This CL imports the windows crosvm perfetto library fork into the
crosvm/perfetto repository. This repository fork is only temporary until
we can converge on a stable C API from the upstream perfetto project.

Another change that this CL has over the normal perfetto repository is
that we removed all the OWNERS files since they conflicted with the
crosvm root OWNERS group.

As per this CL, it currently does not build on Linux due to
Windows-specific dependencies but there is a follow up CL to fix it.

BUG=b:277837670
TEST=None

Change-Id: I2260da8e4e254d3e7639b23d4be424cab31be4d5
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/perfetto/+/4669639
Reviewed-by: Noah Gold <nkgold@google.com>
Auto-Submit: Morg <morg@google.com>